LayoutTests/svg/path-arc-invalid.html

<script>
if (window.layoutTestController)
    layoutTestController.dumpAsText();
</script>
<svg>
<path d="M0 0 A0 0 0 0 0 0 0 0 0 0"/>
</svg>
If this text is visible the malformed SVG path is handled correctly.
